Move ontology related classes to modular files
