+++ encord/objects/utils.py
@@ -1,0 +1,2 @@
+def _decode_nested_uid(nested_uid: list) -> str:
+    return ".".join([str(uid) for uid in nested_uid])
